% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.simulation.R
\name{data.simulation}
\alias{data.simulation}
\title{spCV: Generalized Spatially Varying Coefficient Modeling  (GSVCM) for the
detection of spatial expression patterns for large spatial transcriptomic
studies.}
\usage{
data.simulation(gene.size, k, coords, Z)
}
\arguments{
\item{gene.size}{Number of genes to simulate for each simulation setting.}

\item{k}{Number of cells to be randomly generated from the *coords*.}

\item{coords}{A matrix of all the cell locations.}

\item{Z}{A matrix of all the covariates.}
}
\value{
\item{sample}{Randomly simulated gene expression level.}
\item{S}{A matrix of the random-sampled locations.}
\item{Z}{A matrix of the random-sampled covariates.}
}
\description{
This function generates datasets for simulation studies.
}
