% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.spVC.R
\name{eval.spVC}
\alias{eval.spVC}
\title{spVC: Generalized Spatially Varying Coefficient Modeling  (GSVCM) for the
detection of spatial expression patterns for large spatial transcriptomic
studies.}
\usage{
eval.spVC(coef.gamma, S, V, Tr, center)
}
\arguments{
\item{coef.gamma}{matrix of spline coefficients; each column represents one
coefficients for one spline function.}

\item{S}{A matrix \code{n} by two of locations to be evaluated.}

\item{V}{The \code{nV} by two matrix of verities of a triangulation,
where \code{nV} is the number of vertices. Each row is the coordinates
for a vertex.}

\item{Tr}{The triangulation matrix of dimension \code{nTr} by three,
where \code{nTr} is the number of triangles in the triangulation.}

\item{center}{A vector of estimated mean of spline basis functions.}

\item{coef}{Estimated coefficient matrix of gamma functions from GSVCMs.}
}
\value{
\item{gamma.value}{estimated gamma functions at points \code{S}}
}
\description{
This function generates estimated gamma functions at given locations
}
