% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locTV.R
\name{locTV}
\alias{locTV}
\title{Calculate local energy matrix}
\usage{
locTV(V1, V2, V3, Mat, d)
}
\arguments{
\item{V1, V2, V3}{three vectors of length two indicating the vertices locations of a triangle.}

\item{Mat}{inner coefficients matrix. This arguments could be
calculated by function \code{build}.}

\item{d}{degree of polynomials.}
}
\value{
energy matrix with respect to the triangle \code{(V1,V2,V3)}.
}
\description{
This function computes the local
energy matrix with respect to the triangle \code{(V1,V2,V3)}.
}
