% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.spVC.R
\name{test.spVC}
\alias{test.spVC}
\title{spVC: Generalized Spatially Varying Coefficient Modeling  (GSVCM) for the
detection of spatial expression patterns for large spatial transcriptomic
studies.}
\usage{
test.spVC(
  Y,
  X = NULL,
  S,
  V,
  Tr,
  para.cores = 1,
  scaleX = FALSE,
  subset = 1:nrow(Y),
  p.adjust.method = "BH",
  p.adjust.thresh = 0.05,
  linear.fit = FALSE,
  reduced.only = FALSE,
  twostep = TRUE,
  filter.min.nonzero = 100,
  filter.spot.counts = 100,
  fix.constant = NULL,
  fix.varying = NULL,
  size.factors = NULL
)
}
\arguments{
\item{Y}{A \code{m} by \code{n} matrix of count of RNA-seq reads.
Each row represents one genes, and each column represents one cell.}

\item{X}{A matrix \code{n} by \code{p} of explanatory variables.}

\item{S}{A matrix \code{n} by two of cell locations.}

\item{V}{The \code{nV} by two matrix of verities of a triangulation,
where \code{nV} is the number of vertices. Each row is the coordinates
for a vertex.}

\item{Tr}{The triangulation matrix of dimension \code{nTr} by three,
where \code{nTr} is the number of triangles in the triangulation.}

\item{para.cores}{number of cores in parallel computing.}

\item{scaleX}{an indicator of whether to standardize the covariates.}

\item{subset}{a vector of genes to be tested.}

\item{p.adjust.method}{The p-value adjustment methods.}

\item{p.adjust.thresh}{significant value for adjusted p-value.}

\item{linear.fit}{indicator of whether to fit generalized linear model.}

\item{reduced.only}{indicator of whether to consider the spatially varying
coefficients of the covariates.}

\item{twostep}{indicator of whether to use the two step model estimation
pipeline or directly fit the full model and evaluate the significance of
each individual components in the spVC model.}

\item{filter.min.nonzero}{filter genes whose number of nonzero counts is
larger than \code{filter.min.nonzero}.}

\item{filter.spot.counts}{filter spots whose total gene expression counts is
larger than \code{filter.spot.counts}.}

\item{fix.constant}{subset of covariates only considering the constant effect
in the model.}

\item{fix.varying}{subset of covariates considering the spatially varying
coefficients.}

\item{size.factors}{a vector of given size.factor for each cell. If
\code{size.factor = NULL}, the size.factor is calculated from data.}
}
\value{
\item{results.linear}{A list of spatial pattern testing results for
each gene based on the generalized linear model. We conduct the test for
genes with more than one hundred nonzero counts. Each element contains
p-values of model components and estimate coefficients.}
\item{results.constant}{A list of spatial pattern testing results for
each gene based on Model 1. We conduct the test for genes with more than one hundred nonzero
counts. Each element contains p-values of model components and estimate
coefficients.}
\item{results.varying}{A list of spatial pattern testing results for
each gene based on Model 2. We conduct the test for genes with significant spatial pattern
and explanatory variable. Each element contains p-values of model components
and estimate coefficients.}
\item{BQ2.center.est}{A vector of estimated mean of spline basis functions.}
}
\description{
This function generates the p-value and estimated coefficients of each
component in the GSVCM.
}
